MCMCBenchmarks.jl Documentation

MCMCBenchmarks.jl Documentation

MCMCBenchmarks.jl Documentation

Functions

Index

Functions

MCMCBenchmarks.benchmark — Function.

Runs the benchmarking procedure and returns the results

samplers: a tuple of samplers or a single sampler object

simulate: model simulation function with keyword Nd

Nreps: number of times the benchmark is repeated for each factor combination

source

Index

MCMCBenchmarks.benchmark