-

Sampler Structs

Each sampler is associated with a MCMC sampler struct, which is a subtype of MCMCSampler. In MCMCBenchmarks, we define subtypes of MCMCSampler for three popular NUTS MCMC samplers in Julia: CmdStan, AdvancedHMC via Turing, and DynamicHMC.

MCMCBenchmarks.MCMCSampler — Type.

Abstract MCMC sampler type

source

MCMCBenchmarks.CmdStanNUTS — Type.

MCMC sampler struct for CmdStan NUTS

model: model configuration

dir: probject directory

source

MCMCBenchmarks.AHMCNUTS — Type.

MCMC sampler struct for AdvancedHMC NUTS

model: model function that accepts data

config: sampler configution settings

source

MCMCBenchmarks.DHMCNUTS — Type.

MCMC sampler struct for DynamicHMC NUTS

model: model function that accepts data

config: sampler configution settings

source

Functions

Top-level Benchmark Routines

MCMCBenchmarks.benchmark — Function.

Runs the benchmarking procedure and returns the results

samplers: a tuple of samplers or a single sampler object

simulate: model simulation function with keyword Nd

Nreps: number of times the benchmark is repeated for each factor combination

source

MCMCBenchmarks.pbenchmark — Function.

Runs the benchmarking procedure defined in benchmark in parallel and returns the results

samplers: a tuple of samplers or a single sampler object

simulate: model simulation function with keyword Nd

Nreps: number of times the benchmark is repeated for each factor combination

source

MCMCBenchmarks.benchmark! — Function.

Primary function that performs mcmc benchmark repeatedly on a set of samplers and records the results.

'sampler': tuple of sampler objects

results: DataFrame containing benchmark results

csr̂: cross sampler r̂

simulate: data generating function

Nreps: number of repetitions for a given set of simulation parameters. Default = 100

kwargs: optional keyword arguments that are passed to modifyConfig!, updateResults! and

runSampler, providing flexibility in benchmark simulations.

source

Overloaded Benchmark Functions

MCMCBenchmarks.modifyConfig! — Function.

Modifies MCMC sampler configuration, including number of samples, target acceptance rate and others depending on the specific sampler.

s: sampler object

Nsamples: total number of MCMC samples

Nadapt: number of adaption samples during warm up

delta: target acceptance rate.

source

MCMCBenchmarks.updateResults! — Function.

Update the results DataFrame on each iteration

s: MCMC sampler object

performance: includes MCMC Chain, execution time, and memory measurements

results: DataFrame containing benchmark results

source

MCMCBenchmarks.runSampler — Function.

Extracts model and configuration from sampler object and performs parameter estimation

s: sampler object

data: data for benchmarking

source

Helper Functions

MCMCBenchmarks.addPerformance! — Function.

Adds performance metrics to benchmark results, which include runtime, memory allocations in MB, garbage collection time, percent of time spent in garbage collection, and the number of memory allocations

df: the dataframe to which performance metrics are added

p: a collection of performance metrics including run time,

memory allocations and garbage collection time

source

MCMCBenchmarks.addKW! — Function.

Adds keyword arguments to the results DataFrame

df: DataFrame containing benchmark results for single iteration

kwargs: keyword arguments

source

MCMCBenchmarks.addChainSummary! — Function.

Adds chain summary (e.g. rhat,ess) to newDF for each parameter.

newDF: dataframe that collects results on an iteration

chn: chain for given iteration

df: df of chain results

col: name of column

e.g. If col = :ess, and parameters are mu and sigma, the new columns will be muess and sigmaess and will contain their respective ess values

source

addHPD!

addMeans!

Plotting

MCMCBenchmarks.plotdensity — Function.

Plots a desnity of a distribution for a selected metric (e.g. effective sample size)

df: dataframe of results

metric: name of metric, such as :ess for effective sample size

group: a tuple of grouping factors, e.g. (:sampler,:Nd)

save: save=true saves each plot

figfmt: figure format

dir: directory of saved plot. Default pwd.

source

MCMCBenchmarks.plotsummary — Function.

Plots a summary of a metric computed according to a function, func. func defaults to the mean.

df: dataframe of results

xvar: variable assigned to x-axis

metric: name of metric, such as :ess for effective sample size

group: a tuple of grouping factors, e.g. (:sampler,:Nd)

save: save=true saves each plot

figfmt: figure format

func: a function used to summarize results. Default mean

dir: directory of saved plot. Default pwd.

source

MCMCBenchmarks.plotscatter — Function.

Generates a scatter plot to visualize the relationship between two benchmarking metrics.

df: dataframe of results

xvar: variable assigned to x-axis

metric: name of metric, such as :ess for effective sample size

group: a tuple of grouping factors, e.g. (:sampler,:Nd)

save: save=true saves each plot

figfmt: figure format

func: a function used to summarize results. Default mean

dir: directory of saved plot. Default pwd.

source

MCMCBenchmarks.plotrecovery — Function.

Generates a plot to compare the estimated parameters to true parameter values used to generate simulated data.

df: dataframe of results

parms: namedtuple of parameter names and true values

group: a tuple of grouping factors, e.g. (:sampler,:Nd)

save: save=true saves each plot

figfmt: figure format

func: a function used to summarize results. Default mean

dir: directory of saved plot. Default pwd.

source

Results DataFrame