Functions

Sampler Structs

Each sampler is associated with a MCMC sampler struct, which is a subtype of MCMCSampler. In MCMCBenchmarks, we define subtypes of MCMCSampler for three popular NUTS MCMC samplers in Julia: CmdStan, AdvancedHMC via Turing, and DynamicHMC.

MCMCBenchmarks.MCMCSampler — Type.

Abstract MCMC sampler type

source

MCMCBenchmarks.CmdStanNUTS — Type.

MCMC sampler struct for CmdStan NUTS

model: model configuration

dir: probject directory

id: a unique identifier for each instance of CmdStan in parallel applications

name: a unique identifer given to each sampler

source

MCMCBenchmarks.AHMCNUTS — Type.

MCMC sampler struct for AdvancedHMC NUTS

model: model function that accepts data

config: sampler configution settings

name: a unique identifer given to each sampler

source

MCMCBenchmarks.DHMCNUTS — Type.

MCMC sampler struct for DynamicHMC NUTS

model: model function that accepts data

config: sampler configution settings

name: a unique identifer given to each sampler

source

Functions

Top-level Benchmark Routines

MCMCBenchmarks.benchmark — Function.

Runs the benchmarking procedure and returns the results

samplers: a tuple of samplers or a single sampler object

simulate: model simulation function with keyword Nd

Nreps: number of times the benchmark is repeated for each factor combination

source

julia benchmark!(samplers,results,csr̂,simulate,Nreps,chains;kwargs...)

benchmark!(sampler::T,results,csr̂,simulate,Nreps,chains;kwargs...) where {T<:MCMCSampler}

MCMCBenchmarks.pbenchmark — Function.

Runs the benchmarking procedure defined in benchmark in parallel and returns the results

samplers: a tuple of samplers or a single sampler object

simulate: model simulation function with keyword Nd

Nreps: number of times the benchmark is repeated for each factor combination

source

julia pbenchmark(samplers,simulate,Nreps;kwargs...)

MCMCBenchmarks.benchmark! — Function.

Primary function that performs mcmc benchmark repeatedly on a set of samplers and records the results.

'sampler': tuple of sampler objects

results: DataFrame containing benchmark results

csr̂: cross sampler r̂

simulate: data generating function

Nreps: number of repetitions for a given set of simulation parameters. Default = 100

kwargs: optional keyword arguments that are passed to modifyConfig!, updateResults! and

runSampler, providing flexibility in benchmark simulations.

source

julia benchmark!(samplers,results,csr̂,simulate,Nreps,chains;kwargs...)

Overloaded Benchmark Functions

MCMCBenchmarks.modifyConfig! — Function.

Modifies MCMC sampler configuration, including number of samples, target acceptance rate and others depending on the specific sampler.

s: sampler object

Nsamples: total number of MCMC samples

Nadapt: number of adaption samples during warm up

delta: target acceptance rate.

source

julia modifyConfig!(s::AHMCNUTS;Nsamples,Nadapt,delta,kwargs...)

MCMCBenchmarks.updateResults! — Function.

Update the results DataFrame on each iteration

s: MCMC sampler object

performance: includes MCMC Chain, execution time, and memory measurements

results: DataFrame containing benchmark results

source

julia updateResults!(s::CmdStanNUTS,performance,results;kwargs...)

MCMCBenchmarks.runSampler — Function.

Extracts model and configuration from sampler object and performs parameter estimation

s: sampler object

data: data for benchmarking

Nchains: number of chains ran serially. Default = 1

source

julia runSampler(s::AHMCNUTS,data;kwargs...)

Helper Functions

MCMCBenchmarks.addPerformance! — Function.

Adds performance metrics to benchmark results, which include runtime, memory allocations in MB, garbage collection time, percent of time spent in garbage collection, and the number of memory allocations

df: the dataframe to which performance metrics are added

p: a collection of performance metrics including run time,

memory allocations and garbage collection time

source

julia addPerformance!(df,p)

MCMCBenchmarks.addKW! — Function.

Adds keyword arguments to the results DataFrame

df: DataFrame containing benchmark results for single iteration

kwargs: keyword arguments

source

julia addKW!(df;kwargs...)

MCMCBenchmarks.addChainSummary! — Function.

Adds chain summary (e.g. rhat,ess) to newDF for each parameter.

newDF: dataframe that collects results on an iteration

chn: chain for given iteration

df: df of chain results

col: name of column

e.g. If col = :ess, and parameters are mu and sigma, the new columns will be muess and sigmaess and will contain their respective ess values

source

julia addChainSummary!(newDF,chn,df,col)

addHPD!

julia addHPD!(newDF,chain)

addMeans!

julia addMeans!(newDF,df)

Plotting

MCMCBenchmarks.plotdensity — Function.

Plots a desnity of a distribution for a selected metric (e.g. effective sample size)

df: dataframe of results

metric: name of metric, such as :ess for effective sample size

group: a tuple of grouping factors, e.g. (:sampler,:Nd)

save: save=true saves each plot

figfmt: figure format

dir: directory of saved plot. Default pwd.

source

julia plotdensity(df::DataFrame,metric::Symbol,group=(:sampler,);save=false, figfmt="pdf",dir="",options...)

MCMCBenchmarks.plotsummary — Function.

Plots a summary of a metric computed according to a function, func. func defaults to the mean.

df: dataframe of results

xvar: variable assigned to x-axis

metric: name of metric, such as :ess for effective sample size

group: a tuple of grouping factors, e.g. (:sampler,:Nd)

save: save=true saves each plot

figfmt: figure format

func: a function used to summarize results. Default mean

dir: directory of saved plot. Default pwd.

source

julia plotsummary(df::DataFrame,xvar::Symbol,metric::Symbol,group=(:sampler,);save=false, figfmt="pdf",func=mean,dir="",options...)

MCMCBenchmarks.plotscatter — Function.

Generates a scatter plot to visualize the relationship between two benchmarking metrics.

df: dataframe of results

xvar: variable assigned to x-axis

metric: name of metric, such as :ess for effective sample size

group: a tuple of grouping factors, e.g. (:sampler,:Nd)

save: save=true saves each plot

figfmt: figure format

func: a function used to summarize results. Default mean

dir: directory of saved plot. Default pwd.

source

julia plotscatter(df::DataFrame,xvar::Symbol,metric::Symbol,group=(:sampler,);save=false, figfmt="pdf",func=mean,dir="",options...)

MCMCBenchmarks.plotrecovery — Function.

Generates a plot to compare the estimated parameters to true parameter values used to generate simulated data.

df: dataframe of results

parms: namedtuple of parameter names and true values

group: a tuple of grouping factors, e.g. (:sampler,:Nd)

save: save=true saves each plot

figfmt: figure format

func: a function used to summarize results. Default mean

dir: directory of saved plot. Default pwd.

source

julia plotrecovery(df::DataFrame,parms,group=(:sampler,);save=false, figfmt="pdf",dir="",options...)

$$`

Results DataFrame